requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
playwright>=1.40.0
reportlab>=4.0.0
numpy>=1.26.0
scikit-learn>=1.5.0
requests>=2.28.0
psycopg[binary]>=3.2.1
